Revert "Export getMemes function"
This reverts commit 4da5f57e79d4ddaeb8e695360dd97a47e5e4a44a.
